# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Documentation",
    "description": "RESTful API Documentation - User, Role, Category Management System",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users",
        "description": "Retrieve a list of all users with their roles",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/add": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Add a new user",
        "description": "Create a new user account (Admin only)",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "newuser@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "SecurePass123"
                },
                "first_name": {
                  "type": "string",
                  "example": "John"
                },
                "last_name": {
                  "type": "string",
                  "example": "Doe"
                },
                "phone_number": {
                  "type": "string",
                  "example": "+905551234567"
                },
                "roles": {
                  "type": "array",
                  "example": [
                    "507f1f77bcf86cd799439012"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User created successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/update": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Update an existing user",
        "description": "Update user information and roles",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User update information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "507f1f77bcf86cd799439011"
                },
                "first_name": {
                  "type": "string",
                  "example": "John"
                },
                "last_name": {
                  "type": "string",
                  "example": "Doe"
                },
                "phone_number": {
                  "type": "string",
                  "example": "+905551234567"
                },
                "is_active": {
                  "type": "boolean",
                  "example": true
                },
                "password": {
                  "type": "string",
                  "example": "NewPassword123"
                },
                "roles": {
                  "type": "array",
                  "example": [
                    "507f1f77bcf86cd799439012"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/delete": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a user",
        "description": "Delete a user and their role associations",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User ID to delete",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "507f1f77bcf86cd799439011"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "User login",
        "description": "Authenticate user and receive JWT token",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Login credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful"
          }
        }
      }
    },
    "/api/users/export": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Export users to Excel",
        "description": "Download all users as an Excel file",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Excel file download"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register first user (Initial setup)",
        "description": "Register the first super admin user. Only works if no users exist.",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User registration information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "admin@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "AdminPass123"
                },
                "first_name": {
                  "type": "string",
                  "example": "Admin"
                },
                "last_name": {
                  "type": "string",
                  "example": "User"
                },
                "phone_number": {
                  "type": "string",
                  "example": "+905551234567"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "System initialized successfully"
          }
        }
      }
    },
...
    "/api/routes/": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload a file",
        "description": "Upload a single file to the server",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "file",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "File to upload"
          }
        ],
        "responses": {
          "200": {
            "description": "File uploaded successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}